Add `DateSubtractRequest` class to `models/date_subtract_request.py` and update imports in `python-use-example.py`
* **Add `DateSubtractRequest` class:**
  - Create `models/date_subtract_request.py` file.
  - Define `DateSubtractRequest` class with fields `later_date`, `earlier_date`, and `reason`.
  - Set `extra` configuration to `pydantic.Extra.forbid`.

* **Update imports in `python-use-example.py`:**
  - Import `DateSubtractRequest` class from `models/date_subtract_request.py`.
  - Remove `DateSubtractRequest` class definition from `python-use-example.py`.
